Create program to change text colour.
To print a message in color we use cprintf(). while using cprintf() escape sequences like \n and \t does not work. so we call gotoxy() function to move the pointer to the desired place. gotoxy() takes two parameters i,e one is rows and the other is columns. if you declare gotoxy(2,1); then the pointer goes to the second column, first row and starts printing from there.